% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_integration.R
\name{add_extrapolation_polygons}
\alias{add_extrapolation_polygons}
\title{Add Extrapolation Polygons to Leaflet Map}
\usage{
add_extrapolation_polygons(
  map,
  geometry,
  method = "BUFF",
  color = "blue",
  opacity = 0.5
)
}
\arguments{
\item{map}{Leaflet map object}

\item{geometry}{Spatial geometry object}

\item{method}{Method name ("BUFF", "MPC", "MST")}

\item{color}{Color for polygons}

\item{opacity}{Opacity of polygons}
}
\value{
Updated leaflet map
}
\description{
Add extrapolation polygons to an existing leaflet map
}
