% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stats.R
\name{calc_aic_aicc}
\alias{calc_aic_aicc}
\title{Calculate AIC and AICc}
\usage{
calc_aic_aicc(lnl, k, n)
}
\arguments{
\item{lnl}{Log-likelihood value}

\item{k}{Number of parameters}

\item{n}{Sample size (number of areas)}
}
\value{
List with aic and aicc values
}
\description{
Calculate Akaike Information Criterion and corrected AIC
}
