% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_aicc}
\alias{calc_aicc}
\title{Calculate AICc}
\usage{
calc_aicc(lnl, k, n)
}
\arguments{
\item{lnl}{Log-likelihood value}

\item{k}{Number of parameters}

\item{n}{Sample size}
}
\value{
AICc value
}
\description{
Calculate Akaike Information Criterion corrected for small samples
}
\keyword{internal}
