% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stats.R
\name{calc_akaike_weights}
\alias{calc_akaike_weights}
\title{Calculate Akaike Weights}
\usage{
calc_akaike_weights(aicc)
}
\arguments{
\item{aicc}{Vector of AICc values}
}
\value{
Vector of Akaike weights
}
\description{
Calculate Akaike weights for model comparison
}
