% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stats.R
\name{calc_delta_aic}
\alias{calc_delta_aic}
\title{Calculate Delta AIC and Delta AICc}
\usage{
calc_delta_aic(aic, aicc)
}
\arguments{
\item{aic}{Vector of AIC values}

\item{aicc}{Vector of AICc values}
}
\value{
List with delta_aic and delta_aicc vectors
}
\description{
Calculate differences in AIC and AICc relative to the best model
}
