% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRange_convexHull.R
\name{calcRange_convexHull}
\alias{calcRange_convexHull}
\title{calcRange_convexHull}
\usage{
calcRange_convexHull(xy, shape_file, resol)
}
\arguments{
\item{xy}{object produced by the function \code{\link[PanBioGeo]{singleton_to_data_frame}} or
a table with occurrence points of one or more taxa with columns 'spp', 'long' and 'lat' in decimal degrees}

\item{shape_file}{a polygon of some place already with datum to be used}

\item{resol}{a vector with resolution of quadrats}
}
\value{
Presence/absence matrix based on quadrats - with user-chosen resolution - built
within each minimum convex polygon and a raster with richness values, based on polygon overlap.
}
\description{
Extrapolates distribution by generating minimum convex polygons from taxon occurrence points,
and produces a presence/absence matrix by cell.
}
\details{
This function promotes the generalization of distributions, from minimum convex polygons converted
into quadrats with user-chosen resolution. An 'out_MCP' directory is built and .shp files ('MCP_' for
convex hulls and 'pointshape_' for points), .tif (with richness) and .txt (the binary matrix)
with the polygons and occurrence points of taxa are placed in it.
}
\examples{
library(devtools)
library(tidyverse)
library(lwgeom)
library(tmap)
library(viridis)
library(readr)
library(sf)
library(phytools)
library(terra)

# load example data (coordinates and tree)
data(lycipta)

# The columns are in the following order: species, longitude, latitude
lycipta.coords <- lycipta$coordinates
head(lycipta.coords, 20)

# shape file of South America
lycipta.asul <- lycipta$polygon

# removal of singletons: the columns should be in the following order: species, latitude,
# longitude (without a tree)
resul <- singleton_to_data_frame_without_tree(spp = lycipta.coords$spp,
lat = lycipta.coords$Lat, long = lycipta.coords$Long)

# important to change column names to long and lat
colnames(resul$data_df) <- c('species', 'long', 'lat')
rang1 <- calcRange_convexHull(xy = resul$data_df, shape_file = lycipta.asul,
                               resolut = c(1, 1)) # 1 degree quadrats
rang1$pres_abs # presence and absence matrix
rang1$geometry # raster with values

# plot raster
cores <- viridis(n = 15, alpha = 0.5, option = 'B')
plot(lycipta.asul)
plot(rang1$geometry, col = cores, add = TRUE)

# adding points
nomes <- unique(resul$data_df$species)
points_shapes_n <- list()
pasta <- 'pointshape_'
for(k in 1:length(nomes)){
  if(resul[k] < 3){ # ensuring we keep only taxa with three or more records
      next
  }

  points_shapes_n[[k]] <- vect(paste0('out_MCP/pointshape_', names(resul)[k], '.shp'))
}
points_shapes_n <- points_shapes_n[-which(sapply(points_shapes_n, is.null))]

# plotting shapefiles on the map
for(i in 1:length(points_shapes_n)){
  plot(points_shapes_n[[i]], add = TRUE, col = cores)
}

# adding all minimum convex polygon shapefiles
MCP_shapes_n <- list()
pasta <- 'MCP_'
for(k in 1:length(nomes)){
  if(resul[k] < 3){ # ensuring we keep only taxa with three or more records
      next
  }
  MCP_shapes_n[[k]] <- vect(paste0('out_MCP/MCP_', names(resul)[k], '.shp'))
}
MCP_shapes_n <- MCP_shapes_n[-which(sapply(MCP_shapes_n, is.null))]

# plotting shapefiles on the map
for(i in 1:length(MCP_shapes_n)){
  plot(MCP_shapes_n[[i]], add = TRUE, col = cores)
}
#######################################################################################
# -------- adding user-chosen polygon (e.g., E. L. imitator) ----------
plot(lycipta.asul)
# Search by part of the name
v_imitator_MCP <- MCP_shapes_n[[grep("E_L_imitator", sapply(MCP_shapes_n, sources))[1]]] # MCP
v_imitator_points <- points_shapes_n[[grep("E_L_imitator", sapply(points_shapes_n, sources))[1]]] # points

plot(neo)
plot(v_imitator_MCP, add = TRUE, col = 'grey')
plot(v_imitator_points, add = TRUE, col = 'green')
########################################################################################
# Integration with extrapolation_plot_from_shapeFiles
rang1 <- calcRange_convexHull(xy = resul$data_df, shape_file = lycipta.asul,
                               resol = c(1, 1)) # quadrats
plot(neo, axes = TRUE, las = 1, main = "Species MCPs")
extrapolation_plot_from_shapeFiles(c("E_L_triangulator", "E_L_aceratos"),
             directory = "out_MCP", file_type = "MCP")
########################################################################################
# Doing a PAE-PCE analysis using convex hulls...
rang1 <- calcRange_convexHull(xy = resul$data_df, shape_file = lycipta.asul,
                               resol = c(5, 5)) # quadrats
pae_convexHulls <- pae_pce(preabsMat = rang1$pres_abs, shapeFile = lycipta.asul,
                           resol = c(5, 5), gridView = TRUE, labelGrid = TRUE,
                           nonHomoplasticSpeciesList = TRUE, N = 10, sobrepo = FALSE)

}
\references{
Castillo-Garcia, C.F., Morrone, J.J., Salgado-Ugarte, I.H. & D. Espinosa, 2025. Panbiotracks: software for track analysis.
\emph{Revista Mexicana de Biodiversidad} \strong{96}: e965429.

Page, R.D.M., 1987. Graphs and generalized tracks: quantifying Croizat's panbiogeography.
\emph{Systematic Zoology} \strong{36}: 1.
}
\seealso{
\code{\link[phytools]{phytools}} for phylogenetic trees,
\code{\link[terra]{terra}} for spatial data handling,
\code{\link[PanBioGeo]{MST_node}} for the production of presence-absence matrices.
\code{\link[PanBioGeo]{pae_pce}} for the production of generalized tracks.
\code{\link[PanBioGeo]{extrapolation_plot_from_shapeFiles}} for plotting species distributions.
}
\author{
Jose Ricardo Inacio Ribeiro \email{joseribeiro@unipampa.edu.br}

Augusto Ferrari \email{ferrariaugusto@gmail.com}
}
