% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_integration.R
\name{create_biogeobears_file}
\alias{create_biogeobears_file}
\title{Create BioGeoBEARS Geography File}
\usage{
create_biogeobears_file(pres_abs, method = "BUFF", output_dir = ".")
}
\arguments{
\item{pres_abs}{Presence-absence matrix}

\item{method}{Method name ("BUFF", "MPC", "MST")}

\item{output_dir}{Output directory}
}
\value{
Path to created file
}
\description{
Convert presence-absence matrix to BioGeoBEARS .data format
}
