% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_integration.R
\name{extract_pres_abs}
\alias{extract_pres_abs}
\title{Extract Presence-Absence Matrix}
\usage{
extract_pres_abs(result)
}
\arguments{
\item{result}{Result from range extrapolation function}
}
\value{
Presence-absence matrix
}
\description{
Extract pres_abs from range function results
}
