% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_results}
\alias{format_results}
\title{Format Model Results}
\usage{
format_results(results)
}
\arguments{
\item{results}{List of BioGeoBEARS results}
}
\value{
Data frame with formatted results
}
\description{
Format BioGeoBEARS results for display
}
\keyword{internal}
