% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output_files.R
\name{get_available_output_dirs}
\alias{get_available_output_dirs}
\title{Get list of available output directories}
\usage{
get_available_output_dirs()
}
\value{
Character vector of directories that exist
}
\description{
Get list of available output directories
}
