% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output_files.R
\name{get_raster_names}
\alias{get_raster_names}
\title{Get list of raster names from output directories}
\usage{
get_raster_names()
}
\value{
Character vector of raster names
}
\description{
Get list of raster names from output directories
}
