% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output_files.R
\name{get_shapefile_names}
\alias{get_shapefile_names}
\title{Get list of shapefile names from output directories}
\usage{
get_shapefile_names()
}
\value{
Character vector of shapefile names
}
\description{
Get list of shapefile names from output directories
}
