% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{get_tree_stats}
\alias{get_tree_stats}
\title{Get Tree Statistics}
\usage{
get_tree_stats(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
List with tree statistics
}
\description{
Calculate summary statistics for a phylogenetic tree
}
