% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{has_branch_lengths}
\alias{has_branch_lengths}
\title{Tree Utility Functions for PAE-PCE Analysis}
\usage{
has_branch_lengths(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
Logical, TRUE if tree has branch lengths
}
\description{
Functions to prepare phylogenetic trees for PAE-PCE analysis
}
\details{
Check if Tree has Branch Lengths

Determine if a phylogenetic tree has branch length information
}
