% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{is_ultrametric}
\alias{is_ultrametric}
\title{Check if Tree is Ultrametric}
\usage{
is_ultrametric(tree)
}
\arguments{
\item{tree}{phylo object}
}
\value{
Logical, TRUE if tree is ultrametric
}
\description{
Determine if a tree is ultrametric (all tips equidistant from root)
}
