% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output_files.R
\name{load_output_shapefiles}
\alias{load_output_shapefiles}
\title{Load Shapefiles from Output Directories}
\usage{
load_output_shapefiles()
}
\value{
A list containing:
\itemize{
\item shapefiles: named list of loaded shapefiles (terra objects)
\item rasters: named list of loaded rasters
\item directories: list of directories searched
}
}
\description{
Searches for and loads all shapefiles from output directories created by
range extrapolation functions (out_buffers, out_MCP, out_MST, etc.)
}
