% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_integration.R
\name{load_phylogenetic_tree}
\alias{load_phylogenetic_tree}
\title{Range Extrapolation Integration Functions}
\usage{
load_phylogenetic_tree(tree_file)
}
\arguments{
\item{tree_file}{Path to tree file}
}
\value{
phylo object
}
\description{
Wrapper functions to integrate actual range extrapolation methods into the Shiny app
}
\details{
Load Phylogenetic Tree

Load a phylogenetic tree from file (Newick or Nexus format)
}
