% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utilities.R
\name{make_ultrametric}
\alias{make_ultrametric}
\title{Convert Tree to Ultrametric}
\usage{
make_ultrametric(tree, method = "equal")
}
\arguments{
\item{tree}{phylo object with branch lengths}

\item{method}{Method for ultrametricization:
\itemize{
\item "equal": All branch lengths become 1 (equal length tree)
\item "scale": Scale existing branch lengths proportionally
\item "penalized": Use penalized least squares (requires phytools)
}}
}
\value{
Ultrametric phylo object
}
\description{
Convert a non-ultrametric tree to ultrametric by scaling branch lengths
so all tips are equidistant from the root. This is optional and can be done
by user request before PAE-PCE analysis.
}
