% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST_node.R
\name{MST_node}
\alias{MST_node}
\title{MST_node}
\usage{
MST_node(
  coordin,
  tree = NULL,
  shape_file,
  resol,
  seeres = FALSE,
  sobrepo = FALSE,
  mintreeall = FALSE,
  caption = TRUE,
  nodes = NULL,
  taxon = NULL,
  pol = FALSE,
  transp = 0.5,
  seephylog = FALSE,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{coordin}{coordinates - in decimal degrees - with the following header: spp, lat,
long (all lowercase)}

\item{tree}{a phylogenetic tree in Newick format}

\item{shape_file}{a polygon of some place already with datum}

\item{resol}{a vector with resolution of quadrats}

\item{seeres}{logical value that enables seeing the resolution of quadrats on the map}

\item{sobrepo}{logical value that enables plotting on a map a new minimum spanning trees (MSTs) have already
been produced}

\item{mintreeall}{logical value that enables plotting a single minimum spanning tree (MST) that represents
all species}

\item{caption}{logical value that enables the labeling of a plot based on the specified species}

\item{nodes}{a vector with the node(s) of a tree that allows for plotting a single minimum spanning tree (MST)
representing all species related to that (those) node(s) (optional)}

\item{taxon}{a vector containing the name(s) of terminal taxon (taxa) for creating a minimum spanning tree (MST)
for each of them}

\item{transp}{a value (from 0 to 1) representing the degree of transparency of a polygon}

\item{seephylog}{a logical option to visualize the phylogeny with nodes on the map (optional)}

\item{...}{optional plot arguments}
}
\value{
Matrix with a number of rows equal to the number of cells of the polygon, determined by
the resol parameter, and with a number of columns equal to the number of taxa.
}
\description{
Estimates the minimum spanning tree (MST) for a given taxon, taxa of a given node or nodes,
both internal and terminal. It also estimates the MST of a given species regardless of its presence
on a tree.
}
\details{
This function produces individual tracks, which are the basic unit of a panbiogeographical
analysis (Page, 1987; Castillo-Garcia \emph{et al.}, 2025). It defines a minimum spanning tree,
which is a graph with \emph{n} localities connected through \emph{n}-1 edges, producing a shortest
total length graph.
}
\examples{
# load example data (coordinates and tree)
library(phytools)
library(terra)
library(sf)
data(lycipta)

# The columns are in the following order: species, longitude, latitude
lycipta.coords <- lycipta$coordinates
head(lycipta.coords, 20)

# phylogenetic tree
lycipta.tree <- lycipta$phylogenetic_tree

# removal of singletons: the columns should be in the following order: species, latitude,
# longitude (with a tree)
resul <- singleton_to_data_frame(data = lycipta.coords[, c(1, 3, 2)], phylogeny = lycipta.tree)

#' # removal of singletons: the columns should be in the following order: species, latitude,
# longitude (without a tree)
resul <- singleton_to_data_frame(data = lycipta.coords[, c(1, 3, 2)])
###############################################################################################
# Download South America countries (low or high resolutions)
sa_countries <- rnaturalearth::ne_countries(continent = "South America", scale = "medium",
  returnclass = "sf")

sa_countries <- rnaturalearth::ne_countries(continent = "South America", scale = "large",
  returnclass = "sf")

brazil_states <- rnaturalearth::ne_states(country = "Brazil", returnclass = "sf")

# removing Brazil
sa_no_br <- sa_countries[sa_countries$iso_a3 != "BRA", ]

# standardizing columns (keeping geometry column intact)
countries <- sa_no_br[, c("name", "iso_a3")]
names(countries)[1:2] <- c("region_name", "iso_code")
countries$type <- "country"

states <- brazil_states[, c("name", "iso_3166_2")]
names(states)[1:2] <- c("region_name", "iso_code")
states$type <- "state"

st_is_valid(countries)
st_is_valid(states) # there is a problem here!

# Validating gemometries
countries_clean <- st_make_valid(countries)
states_clean <- st_make_valid(states)

# Combining
sa_bins <- rbind(countries_clean, states_clean)

# Convert to terra SpatVector if needed
sa_bins_sv <- terra::vect(sa_bins)


# ---------------------------- with all taxa ------------------------------
# ----------- with a tree and resolution of 10 X 10 degrees ----------
mst_all_taxa <- MST_node(coordin = resul$data_df, shape_file = sa_bins_sv, sobrepo = FALSE, caption = TRUE,
  resol = c(10, 10), tree = resul$treeMod)

# -------- without a tree and resolution of 10 X 10 degrees ----------
mst_all_taxa <- MST_node(coordin = resul$data_df, shape_file = sa_bins_sv, sobrepo = FALSE, caption = TRUE,
  resol = c(10, 10))
#################################################################################################
# ---------------- with terminals: E. l. triangulator ----------------
resul_triangulator <- MST_node(taxon = 'E_L_triangulator', coordin = resul$data_df, shape_file = sa_bins_sv,
  sobrepo = FALSE, caption = TRUE, resol = c(10, 10), tree = resul$treeMod)

# --------------- with terminals: E. l. triangulator + E. l. aceratos
resul_twoTaxa <- MST_node(taxon = c('E_L_triangulator', 'E_L_aceratos'), coordin = resul$data_df,
  shape_file = sa_bins_sv, sobrepo = FALSE, caption = TRUE, resol = c(10, 10), tree = resul$treeMod)
#################################################################################################
# ----------------------------- with internal nodes --------------------------------
# To determine the minimum spanning tree of members in an internal node, we first need to know the number of internal
# nodes in the tree.
 plotTree(resul$treeMod)
 labelnodes(1:(Ntip(resul$treeMod) + resul$treeMod$Nnode), 1:(Ntip(resul$treeMod) + resul$treeMod$Nnode),
  interactive = F, cex = 0.6, circle.exp = 0.4)

# Now that we know the numbers of the internal nodes, we will ask R to calculate the individual tracks of all
# members of any monophyletic group.
# internal nodes: node number 23
resul_node_23 <- MST_node(coordin = resul$data_df, tree = resul$treeMod, shape_file = sa_bins_sv,
 caption = TRUE, resol = c(10, 10), nodes = 23, sobrepo = FALSE)
##################################################################################################
# ------------------- internal nodes + terminal nodes (E_L_imitator) ---------------
resul_node23_imitator <- MST_node(coordin = resulcTree$data_df, tree = resulcTree$treeMod, shape_file = asul,
caption = TRUE, resol = c(10, 10), nodes = 23, taxon = 'E_L_imitator')
##################################################################################################
# the map has been plotted! Now, let's ask R to calculate the individual track of a taxon at a node
resul_node20_sobrepo <- MST_node(coordin = resul$data_df, tree = resul$treeMod, shape_file = asul,
caption = TRUE, resol = c(10, 10), sobrepo = T, nodes = 20)

}
\references{
Castillo-Garcia, C.F., Morrone, J.J., Salgado-Ugarte, I.H. & D. Espinosa, 2025. Panbiotracks: software for track analysis.
\emph{Revista Mexicana de Biodiversidad} \strong{96}: e965429.

Page, R.D.M., 1987. Graphs and generalized tracks: quantifying Croizat's panbiogeography.
\emph{Systematic Zoology} \strong{36}: 1.
}
\seealso{
\code{\link[phytools]{phytools}} for phylogenetic trees,
\code{\link[rnaturalearthdata]{rnaturalearthdata}} for global geographic data,
\code{\link[terra]{terra}} for spatial data handling.
}
\author{
Jose Ricardo Inacio Ribeiro \email{joseribeiro@unipampa.edu.br}

Augusto Ferrari \email{ferrariaugusto@gmail.com}
}
