% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pae_pce.R
\name{pae_pce}
\alias{pae_pce}
\title{pae_pce}
\usage{
pae_pce(
  preabsMat,
  shapeFile,
  resol,
  N = NULL,
  gridView = FALSE,
  labelGrid = FALSE,
  nonHomoplasticSpeciesList = TRUE,
  sobrepo = FALSE,
  random_seed = 123,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  use_hatching = TRUE,
  hatching_angles = NULL
)
}
\arguments{
\item{preabsMat}{the presence-absence matrix from MST_node function}

\item{shapeFile}{a polygon of some place already with datum to be used}

\item{resol}{a vector with resolution of quadrats}

\item{N}{number of iterations in order to get the most parsimonious tree}

\item{gridView}{logical value that enables viewing a grid with a given resolution provided by
the resol argument}

\item{labelGrid}{logical value that enables viewing the numbers of a grid with a given resolution provided by
the resol argument}

\item{nonHomoplasticSpeciesList}{logical value that enables the production of a table (in .txt) showing
the presence of non-homoplastic species and its referred number of grid in a /out directory automatically produced}

\item{sobrepo}{logical value that enables plotting on a map with minimum spanning trees (MSTs) already
produced}

\item{...}{optional plot arguments}
}
\value{
Matrix with homoplastic and non-homoplastic species and their grid numbers associated, as well as a plot
with the polygon indicated by the shapeFile argument and the hatched grids indicating the generalized track
}
\description{
Produces generalized tracks with PAE-PCE, from the presence-absence matrix produced by the MST_node function.
}
\details{
This function produces generalized tracks, which can be used as primary biogeographical homology
hypothesis in Biogeography (Page, 1987; Castillo-Garcia \emph{et al.}, 2025). It is defined as the superimposition
of at least two minimum spanning trees, which are graphs with \emph{n} localities connected through \emph{n}-1 edges, producing shortest
total length graphs.
}
\examples{
# First, determine the minimum spanning trees of all taxa and produce a presence-absence matrix
# ---------------------------- with all taxa ------------------------------
# ----------- with a tree and resolution of 10 X 10 degrees ----------
mst_all_taxa <- MST_node(coordin = resul$data_df, shape_file = lycipta.asul, sobrepo = F, caption = T,
  resol = c(10, 10), tree = resul$treeMod)

# Now, produce the generalized track and a list with grid numbers and the characters (i.e., taxaset)
# Please try to run at least more than one round, with N > 1...
# With a new plot using the polygon of South America (asul)
pae1 <- pae_pce(preabsMat = mst_all_taxa, shapeFile = asul, resol = c(10, 10), gridView = TRUE, labelGrid = TRUE,
   nonHomoplasticSpeciesList = FALSE, N = 10)
pae1 # data frame showing the homoplastic and non-homoplastic species (spp) and their grid numbers (grid_n),
# from the anterior iterations (PAE-PCE)

# Plotting onto a map already plotted with minimum spanning trees
pae2 <- pae_pce(preabsMat = mst_all_taxa, shapeFile = asul, resol = c(10, 10),
               gridView = TRUE, labelGrid = TRUE, nonHomoplasticSpeciesList = TRUE, sobrepo = TRUE,
               N = 10)
pae2

# Doing a zoom
pae3 <- pae_pce(preabsMat = mst_all_taxa, shapeFile = asul, resol = c(10, 10),
               gridView = TRUE, labelGrid = TRUE, nonHomoplasticSpeciesList = TRUE, sobrepo = TRUE,
               N = 10, xmin = min(lycipta.coords$Long) - 5,
               xmax = max(lycipta.coords$Long), ymin = min(lycipta.coords$Lat) - 7,
               ymax = max(lycipta.coords$Lat))
pae3

}
\references{
Castillo-Garcia, C.F., Morrone, J.J., Salgado-Ugarte, I.H. & D. Espinosa, 2025. Panbiotracks: software for track analysis.
\emph{Revista Mexicana de Biodiversidad} \strong{96}: e965429.

Page, R.D.M., 1987. Graphs and generalized tracks: quantifying Croizat's panbiogeography.
\emph{Systematic Zoology} \strong{36}: 1.
}
\seealso{
\code{\link[phytools]{phytools}} for phylogenetic trees,
\code{\link[terra]{terra}} for spatial data handling,
\code{\link[PanBioGeo]{MST_node}} for the production of presence-absence matrices.
}
\author{
Jose Ricardo Inacio Ribeiro \email{joseribeiro@unipampa.edu.br}

Augusto Ferrari \email{ferrariaugusto@gmail.com}
}
