% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stats.R
\name{perform_lrt}
\alias{perform_lrt}
\title{Perform Likelihood Ratio Test}
\usage{
perform_lrt(lnl_simple, lnl_complex, df)
}
\arguments{
\item{lnl_simple}{Log-likelihood of simple model}

\item{lnl_complex}{Log-likelihood of complex model}

\item{df}{Degrees of freedom (difference in number of parameters)}
}
\value{
List with test statistic, p-value, and significance
}
\description{
Test if a complex model is significantly better than a simple model
}
