% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_extrapolation_polygons}
\alias{plot_extrapolation_polygons}
\title{Plot Extrapolation Polygons on Map}
\usage{
plot_extrapolation_polygons(map, geometry, species_names)
}
\arguments{
\item{map}{leaflet map object}

\item{geometry}{Spatial geometry object (SpatVector or sf)}

\item{species_names}{Character vector of species names}
}
\value{
Updated leaflet map object
}
\description{
Add extrapolation polygons (buffers, convex hulls, MST) to a leaflet map
}
