% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNDM.R
\name{process_occurrences}
\alias{process_occurrences}
\title{Process occurrence data}
\usage{
process_occurrences(input_data, separator, header, na_strings, remove_na)
}
\description{
Process occurrence data
}
\keyword{internal}
