% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNDM.R
\name{process_shapefiles}
\alias{process_shapefiles}
\title{Process shapefiles from extrapolated distributions}
\usage{
process_shapefiles(shapefile_dir, resolution, shape_file)
}
\description{
Process shapefiles from extrapolated distributions
}
\keyword{internal}
