% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output_files.R
\name{raster_to_raster_layer}
\alias{raster_to_raster_layer}
\title{Convert Raster to Leaflet-compatible format}
\usage{
raster_to_raster_layer(raster)
}
\arguments{
\item{raster}{A terra raster object}
}
\value{
A RasterLayer object
}
\description{
Converts terra raster to RasterLayer for use with leaflet
}
