% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_output_files.R
\name{shapefile_to_sf}
\alias{shapefile_to_sf}
\title{Convert Shapefile to Leaflet-compatible format}
\usage{
shapefile_to_sf(shapefile)
}
\arguments{
\item{shapefile}{A terra SpatVector object}
}
\value{
An sf object
}
\description{
Converts terra SpatVector to sf for use with leaflet
}
