% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleton_to_data_frame.R
\name{singleton_to_data_frame}
\alias{singleton_to_data_frame}
\title{singleton_to_data_frame}
\usage{
singleton_to_data_frame(data = NULL, phylogeny = NULL)
}
\arguments{
\item{data}{coordinates - in decimal degrees - with the following header: \strong{spp, lat,
long (all lowercase)}}

\item{phylogeny}{a phylogenetic tree in Newick format}
}
\value{
Both a new dataset as a data frame object and a modified tree without singletons.
}
\description{
Removes singletons from a dataset and prunes taxa from a phylogeny.
}
\details{
It can be utilized even if the dataset does not contain any singletons, which may only be discovered later.
It is also applicable in situations where a taxon present in the tree is absent from the dataset, but it is also applicable
in situations where a taxon is absent from the tree and present in the dataset. Finally, it can be used without a phylogeny.
The idea was based on the phytools package (Revell, 2024).
}
\examples{
# load example data (coordinates and tree)
library(phytools)
data(lycipta)

# The columns are in the following order: species (spp), longitude (long), latitude (lat)
lycipta.coords <- lycipta$coordinates
head(lycipta.coords, 20)

# phylogenetic tree
lycipta.tree <- lycipta$phylogenetic_tree

# removal of singletons with a tree
resul <- singleton_to_data_frame(data = lycipta.coords[, c(1, 3, 2)], phylogeny = lycipta.tree)

# removal of singletons without a tree
resul <- singleton_to_data_frame(data = lycipta.coords[, c(1, 3, 2)])
}
\references{
Revell, L.J., 2024. phytools 2.0: an updated R ecosystem for phylogenetic comparative methods (and other
things). \emph{PeerJ} \strong{12}: e16505.
}
\seealso{
\code{\link[phytools]{drop.tip}} for pruning a taxon out of phylogenetic trees.

\code{\link[PanBioGeo]{MST_node}} for estimating the minimum spanning tree (MST).
}
\author{
Jose Ricardo Inacio Ribeiro \email{joseribeiro@unipampa.edu.br}

Augusto Ferrari \email{ferrariaugusto@gmail.com}
}
