% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesGeneTrackPlot.R
\name{speciesGeneTrackPlot}
\alias{speciesGeneTrackPlot}
\title{speciesGeneTrackPlot}
\usage{
speciesGeneTrackPlot(PAE_PCE, resol, spvect, runPaePce)
}
\arguments{
\item{PAE_PCE}{object produced by the function \code{\link[PanBioGeo]{pae_pce}} with the list of species (non-homoplastic
synapomorphies) of each clade found}

\item{resol}{a vector with resolution of quadrats}

\item{spvect}{a polygon of some place already with datum to be used}

\item{runPaePce}{a vector indicating which PAE-PCE run is to be indicated on the map}
}
\value{
Map with hatched and properly numbered grid cells indicating the estimated generalized tracks and the
occurrence points of species (synapomorphies) associated with their MSTs located in the /out directory
}
\description{
Produces a map - to be chosen by the user (in shapefile format) - and plots estimated generalized tracks
with PAE-PCE, in the form of grid cells with the same resolution selected in the analysis.
}
\details{
This function plots the generalized tracks produced by the pae_pce function - one for each run (when applicable) -
for a given set of species, whose occurrence points and MSTs are already available in the /out directory.
The produced generalized track is also found in the /out directory with the name "GeneralizedTrack_#.tif", where #
represents the run number of the PAE-PCE analysis.
}
\examples{
# First, determine the minimum spanning trees of all taxa and produce a presence-absence matrix
# ---------------------------- with all taxa ------------------------------
# ----------- with a tree and resolution of 10 X 10 degrees ---------------
mst_all_taxa <- MST_node(coordin = resul$data_df, shape_file = lycipta.asul, sobrepo = F, caption = T, resol = c(10, 10), tree = resul$treeMod)

# --------------------------------- generalized track ---------------------
# Now, produce the generalized track and a list with grid numbers and the characters (i.e., taxaset)
# Please try to run at least more than one round, with N > 1...
# With a new plot using the polygon of South America (asul)
pae1 <- pae_pce(preabsMat = mst_all_taxa, shapeFile = asul, resol = c(10, 10), gridView = TRUE, labelGrid = TRUE, nonHomoplasticSpeciesList = FALSE, N = 10)
pae1 # data frame showing the homoplastic and non-homoplastic species (spp) and their grid numbers (grid_n), from the anterior iterations (PAE-PCE)

speciesPlot(PAE_PCE = pae1, resol = c(10, 10), spvect = neo, runPaePce = 1)
}
\references{
Castillo-Garcia, C.F., Morrone, J.J., Salgado-Ugarte, I.H. & D. Espinosa, 2025. Panbiotracks: software for track analysis.
\emph{Revista Mexicana de Biodiversidad} \strong{96}: e965429.

Page, R.D.M., 1987. Graphs and generalized tracks: quantifying Croizat's panbiogeography.
\emph{Systematic Zoology} \strong{36}: 1.
}
\seealso{
\code{\link[phytools]{phytools}} for phylogenetic trees,
\code{\link[terra]{terra}} for spatial data handling,
\code{\link[PanBioGeo]{MST_node}} for the production of presence-absence matrices.
\code{\link[PanBioGeo]{pae_pce}} for the production of generalized tracks.
}
\author{
Jose Ricardo Inacio Ribeiro \email{joseribeiro@unipampa.edu.br}

Augusto Ferrari \email{ferrariaugusto@gmail.com}
}
