% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_tree_file}
\alias{validate_tree_file}
\title{Validate Tree File}
\usage{
validate_tree_file(tree_file)
}
\arguments{
\item{tree_file}{Path to tree file}
}
\value{
Logical, TRUE if valid
}
\description{
Check if tree file is valid Newick format
}
\keyword{internal}
